+++ tests/redis.js
@@ -1,93 +1,150 @@
+/* eslint-disable no-underscore-dangle */
+
 'use strict';
 
 const assert = require('assert');
 const sinon = require('sinon');
 
-const redis = require('@redis/client');
+const RedisLib = require('@redis/client');
 
 const Settings = require('@janiscommerce/settings');
+
+const Events = require('@janiscommerce/events');
 
 const Redis = require('../lib/redis');
 
 describe('Redis', () => {
 
+	const originalEnvs = { ...process.env };
+
 	const stubSettings = returns => {
 		sinon.stub(Settings, 'get')
-			.returns(returns);
+			.returns(returns || {});
 	};
 
 	afterEach(() => {
 		sinon.assert.calledOnceWithExactly(Settings.get, 'redis');
 		sinon.restore();
-		// eslint-disable-next-line no-underscore-dangle
+		process.env = { ...originalEnvs };
+
 		Redis._config = undefined;
+		Redis.cleanCluster();
 	});
 
-	it('Should create redis client using settings', async () => {
+	context('When Redis is configured with env vars', () => {
 
-		const clientStub = sinon.stub();
+		it('Should create Redis cluster using env vars REDIS_WRITE_URL', async () => {
 
-		const client = { connect: clientStub };
+			process.env.REDIS_WRITE_URL = 'write.redis.my-service.com';
 
-		sinon.stub(redis, 'createClient')
-			.returns(client);
+			stubSettings();
 
-		stubSettings({ host: 'localhost', port: 6379 });
+			const clusterStub = sinon.stub();
 
-		const createdClient = await Redis.connect();
+			const cluster = { connect: clusterStub };
 
-		assert.deepStrictEqual(createdClient, client);
+			sinon.stub(RedisLib, 'createCluster')
+				.returns(cluster);
 
-		sinon.assert.calledOnceWithExactly(redis.createClient, {
-			socket: {
-				host: 'localhost',
-				port: 6379,
-				reconnectStrategy: sinon.match.func
-			}
-		});
+			const createdCluster = await Redis.connect();
 
-		sinon.assert.calledOnceWithExactly(clientStub);
-	});
+			assert.deepStrictEqual(createdCluster, cluster);
 
-	context('When reconnect strategy is used', () => {
-
-		const client = { connect: () => {} };
-
-		beforeEach(() => {
-			stubSettings({ host: 'localhost', port: 6379 });
-		});
-
-		afterEach(() => {
-			sinon.assert.calledOnceWithExactly(redis.createClient, {
-				socket: {
-					host: 'localhost',
-					port: 6379,
-					reconnectStrategy: sinon.match.func
-				}
+			sinon.assert.calledOnceWithExactly(RedisLib.createCluster, {
+				rootNodes: [{ url: 'write.redis.my-service.com' }],
+				useReplicas: true
 			});
+
+			sinon.assert.calledOnceWithExactly(clusterStub);
 		});
 
-		[1000, 1000, false, false].forEach((expectedResult, retry) => {
+		it('Should create Redis cluster using env vars REDIS_WRITE_URL and REDIS_READ_URL', async () => {
 
-			it(`Should response ${expectedResult} on retry #${retry}`, async () => {
+			process.env.REDIS_WRITE_URL = 'write.redis.my-service.com';
+			process.env.REDIS_READ_URL = 'read.redis.my-service.com';
 
-				sinon.stub(redis, 'createClient')
-					.callsFake(options => {
-						const result = options.socket.reconnectStrategy(retry);
-						assert.deepStrictEqual(result, expectedResult);
+			stubSettings();
 
-						return client;
-					});
+			const clusterStub = sinon.stub();
 
-				await Redis.connect();
+			const cluster = { connect: clusterStub };
 
+			sinon.stub(RedisLib, 'createCluster')
+				.returns(cluster);
+
+			const createdCluster = await Redis.connect();
+
+			assert.deepStrictEqual(createdCluster, cluster);
+
+			sinon.assert.calledOnceWithExactly(RedisLib.createCluster, {
+				rootNodes: [
+					{ url: 'write.redis.my-service.com' },
+					{ url: 'read.redis.my-service.com' }
+				],
+				useReplicas: true
 			});
+
+			sinon.assert.calledOnceWithExactly(clusterStub);
 		});
 	});
 
-	it('Should avoid connection when no settings were found', async () => {
+	context('When Redis is configured with Settings file', () => {
 
-		sinon.spy(redis, 'createClient');
+		it('Should create Redis cluster using old settings', async () => {
+
+			const clusterStub = sinon.stub();
+
+			const cluster = { connect: clusterStub };
+
+			sinon.stub(RedisLib, 'createCluster')
+				.returns(cluster);
+
+			stubSettings({ host: 'redis.my-service.com' });
+
+			const createdCluster = await Redis.connect();
+
+			assert.deepStrictEqual(createdCluster, cluster);
+
+			sinon.assert.calledOnceWithExactly(RedisLib.createCluster, {
+				rootNodes: [{ url: 'redis.my-service.com' }],
+				useReplicas: true
+			});
+
+			sinon.assert.calledOnceWithExactly(clusterStub);
+		});
+	});
+
+	context('Re-using cluster connection', () => {
+
+		it('Should reuse Redis cluster after connect a second time', async () => {
+
+			process.env.REDIS_WRITE_URL = 'write.redis.my-service.com';
+
+			const clusterStub = sinon.stub();
+
+			const cluster = { connect: clusterStub };
+
+			sinon.stub(RedisLib, 'createCluster')
+				.returns(cluster);
+
+			stubSettings();
+
+			const createdCluster = await Redis.connect();
+
+			assert.deepStrictEqual(createdCluster, cluster);
+
+			const otherCluster = await Redis.connect();
+
+			assert.deepStrictEqual(createdCluster, otherCluster);
+
+			sinon.assert.calledOnce(RedisLib.createCluster);
+			sinon.assert.calledOnceWithExactly(clusterStub);
+		});
+	});
+
+	it('Should not connect when no settings nor env vars were configured', async () => {
+
+		sinon.spy(RedisLib, 'createClient');
 
 		sinon.stub(Settings, 'get')
 			.returns();
@@ -96,6 +153,31 @@ describe('Redis', () => {
 
 		assert.deepStrictEqual(connResult, undefined);
 
-		sinon.assert.notCalled(redis.createClient);
+		sinon.assert.notCalled(RedisLib.createClient);
+	});
+
+	it('Should closeConnection after janiscommerce.ended is emitted', async () => {
+
+		process.env.REDIS_WRITE_URL = 'write.redis.my-service.com';
+
+		const clusterStub = sinon.stub();
+		const quitStub = sinon.stub();
+
+		const cluster = {
+			connect: clusterStub,
+			quit: quitStub
+		};
+
+		sinon.stub(RedisLib, 'createCluster')
+			.returns(cluster);
+
+		stubSettings();
+
+		await Redis.connect();
+
+		await Events.emit('janiscommerce.ended');
+		await Events.emit('janiscommerce.ended');
+
+		sinon.assert.calledOnceWithExactly(quitStub);
 	});
 });