+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2023-06-16
+### Added
+- New env variables `REDIS_WRITE_URL` and `REDIS_READ_URL` are used
+- Using Clusters
+- Re-using Cluster connection
+- Quit after `janiscommerce.ended` event was emitted
+
+### Deprecated
+- Connection using `@janiscommerce/settings` and `redis` key
+
 ## [2.0.0] - 2023-05-11
 ### Added
 - New async method `connect()`